+++ var/www/api/getFatcaClients.php
@@ -29,6 +29,7 @@
 require_once dirname(__FILE__).'/../../../config.php'; // copy the provided sample in repository/config-sample.php
 
 require_once ROOT_DB_API.'/lib/MarketflowClient.php';
+require_once ROOT_DB_API.'/lib/BankflowClient.php';
 require_once ROOT_DB_API.'/var/www/api/argsProcessor.php';
 require_once ROOT_IDES_DATA.'/lib/libxml_helpers.php';
 require_once ROOT_IDES_DATA.'/lib/Transmitter.php';
@@ -44,12 +45,29 @@ try {
 	// retrieval from mf db table
 	$mfDb=new MarketflowClient($base,$location);
 	$di=$mfDb->getFatcaClients();
-	if($_GET['shuffle']) $di=array2shuffledLetters($di,array("ResidenceCountry"));
+	if($_GET['shuffle']) $di=array2shuffledLetters($di,array("ResidenceCountry","ENT_COD"));
+	$di3=$mfDb->odbc_fetch_array_array(sprintf("select CLI_COD,CLI_ENT_COD from CLIENT where CLI_ENT_COD in (%s)",
+		implode(",",array_map(function($x) { return sprintf("'%s'",$x["ENT_COD"]); }, $di))),
+		'');
 	$mfDb->disconnect();
 
 	if(count($di)==0) throw new Exception("No data");
 
-	$fca=new Transmitter($di);
+	$di4=array_unique(array_map(function($x) { return $x["CLI_COD"]; }, $di3));
+
+	$bfDb=new BankflowClient($base,$location);
+	$di2=$bfDb->cash(date("Y-m-d"),$di4,"entityid");
+	$bfDb->disconnect();
+
+	foreach($di as $k=>$v) {
+		$t=array_filter($di2,function($x) use($v) { return $x["CLI_ENT_COD"]==$v["ENT_COD"]; });
+		$t=array_values($t);
+		$di[$k]["accounts"]=$t;
+	}
+
+	//$di=array_filter($di,function($x) { return count($x["accounts"])>0; }); // omit clients with no balances (closed)
+
+	$fca=new Transmitter($di,true);
 	$fca->toXml(); # convert to xml 
 
 	if(!$fca->validateXml()) {# validate